pyproject.toml
[project]
name = "mygallery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = []
